requirements.txt.txt
streamlit
youtube-transcript-api
google-generativeai
reportlab
tenacity